refactor: Move some annotation to http package
