+++ src/mg/emberframework/annotation/http/Controller.java
@@ -1,4 +1,4 @@
-package mg.emberframework.annotation;
+package mg.emberframework.annotation.http;
 
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -6,7 +6,5 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface RequestParameter {
-    String value();
-}
+@Target(ElementType.TYPE)
+public @interface Controller {}
