+++ src/mg/emberframework/annotation/http/RequestParameter.java
@@ -1,4 +1,4 @@
-package mg.emberframework.annotation;
+package mg.emberframework.annotation.http;
 
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -6,7 +6,7 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.TYPE, ElementType.METHOD })
-public @interface RequiredRole {
-    String[] values();
+@Target(ElementType.PARAMETER)
+public @interface RequestParameter {
+    String value();
 }
